[PW-6497] - Rename the controllers and update helpers
